+++ ourkitty/src/main/kotlin/com/nyang/ourkitty/domain/chart/dto/DishCountResponseDto.kt
@@ -4,7 +4,7 @@ data class DishCountResponseDto(
     // @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
     val date: Int,
     val batteryAmount: Int,
-    val foodAmount: Double,
+    val foodAmount: Int,
     val catCount: Int,
     val noTnrCount: Int,
 ) {
