+++ ourkitty/src/main/kotlin/com/nyang/ourkitty/domain/chart/dto/DishCountResultDto.kt
@@ -2,7 +2,7 @@ package com.nyang.ourkitty.domain.chart.dto
 
 data class DishCountResultDto(
     val batteryAmountList: List<Int>,
-    val foodAmountList: List<Double>,
+    val foodAmountList: List<Int>,
     val catCountList: List<Int>,
     val noTnrCountList: List<Int>,
 ) {
